% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_dfs.R
\name{combine_rds_pacing_dir_v2}
\alias{combine_rds_pacing_dir_v2}
\title{Combine DF's across pacing directions within a timepoint}
\usage{
combine_rds_pacing_dir_v2(
  id,
  df,
  grouping,
  base_dir = "data/",
  suffix = "_full",
  step,
  timepoint,
  wavefront
)
}
\arguments{
\item{id}{Subject id}

\item{df}{Data frame containing data from the pacing direction that you want to merge}

\item{grouping}{"carto_mesh" or "lv_volt" or other group that you've filtering within}

\item{base_dir}{Directory where you're saving this data to}

\item{suffix}{Any suffix you want to add to the data file name, name looks like step2_t2_rvpaced_full.rds}

\item{step}{What step in your process are you in?}

\item{timepoint}{The timepoint that you're currently working with}

\item{wavefront}{The pacing wavefront that you're currently merging in}
}
\value{
Creates a file in wherever we declared our base_dir
}
\description{
Combine DF's across pacing directions within a timepoint
}
