% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiEAM-package.R
\docType{package}
\name{eiEAM-package}
\alias{eiEAM}
\alias{eiEAM-package}
\title{eiEAM: Electroanatomical Mapping Post-Processing Tools}
\description{
eiEAM - Old Mc'Tschabrunn had a farm, eee eye eee eye mapping techniques emphasizing high-density acquisition has led to EP suffering from "Big Data" challenges. This library includes tools for accurate and reproducible analysis of high-density electroanatomic mapping (EAM) data including nearest neighbors co-registration, mesh reprocessing, and statistical analyses.
}
\author{
\strong{Maintainer}: Jonathan Salas \email{Jsalas424@gmail.com} (\href{https://orcid.org/0009-0001-4508-7336}{ORCID})

}
\keyword{internal}
