% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{fr_nearest_neighbors}
\alias{fr_nearest_neighbors}
\title{fr_nearest_neighbors}
\usage{
fr_nearest_neighbors(training_data, testing_data, neighbors_radius = 1)
}
\arguments{
\item{training_data}{Training data is what you want to project on to (matrix or data frame)}

\item{testing_data}{Testing data is what the data you want to project (matrix or data frame)}

\item{neighbors_radius}{Radius (Euclidean Distance in millimeter) between the training and testing data you want to coregister. Higher numbers result in lower specificity coregistration}
}
\value{
Return a dataframe with fixed radius nearest neighbors
}
\description{
fr_nearest_neighbors
}
\examples{
# Create sample data
training_data <- matrix(seq(from = 1, to = 9), nrow = 3, ncol = 3)
testing_data <- matrix(seq(from = 3, to = 32), ncol = 3)
result <- fr_nearest_neighbors(training_data, testing_data, neighbors_radius = 20)
print(result)
}
