% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{n_nearest_neighbors}
\alias{n_nearest_neighbors}
\title{n_nearest_neighbors}
\usage{
n_nearest_neighbors(training_data, testing_data, n_neighbors = 1)
}
\arguments{
\item{training_data}{Training data is what you want to project on to (matrix or data frame)}

\item{testing_data}{Testing data is what the data you want to project (matrix or data frame)}

\item{n_neighbors}{Number of neighbors you want to identify. Higher numbers result in lower specificity coregistration.}
}
\value{
A dataframe with n nearest neighbors
}
\description{
n_nearest_neighbors
}
\examples{
# Create sample data
training_data <- matrix(seq(from = 1, to = 9), nrow = 3, ncol = 3)
testing_data <- matrix(seq(from = 10, to = 27), ncol = 3)
result <- n_nearest_neighbors(training_data, testing_data, n_neighbors = 1)
print(result)
}
