% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiEAM-package.R
\docType{package}
\name{eiEAM-package}
\alias{eiEAM}
\alias{eiEAM-package}
\title{eiEAM: Electroanatomical Mapping Post-Processing Tools}
\description{
Tools for EAM data analysis including nearest neighbor algorithms, mesh visualization, and statistical utilities.
}
\author{
\strong{Maintainer}: Jonathan Salas \email{Jsalas424@gmail.com} (\href{https://orcid.org/0009-0001-4508-7336}{ORCID})

}
\keyword{internal}
